pyproject.toml
[project]
name = "ai-data-science-template"
version = "0.1.0"
description = "Reusable AI + data science project template"
requires-python = ">=3.10"
dependencies = ["hydra-core>=1.3.2"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.4.0", "black>=24.0.0"]
torch = ["torch>=2.0.0"]
hf = ["transformers>=4.38.0", "accelerate>=0.27.0", "deepspeed>=0.12.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py310"
